Update 1
Added and Removed imports, Also renamed User to UserModel and all references, removed Oauth2 features because it broke consistency aswell as renamed path from "/login" to "/token",as well as reworked content username field to properly return user's name instead of being more than that,added request argument to checkAuthorization async function which returns a string based token, aswell as added a TryExcept block to handle error handling,  added annotations to functions to help with typing hints and ect. , created an async function that will taking in a jwt and try to decipher it to get the supposed user, renamed test async function to protectedRoute, created async functions to logout, retrieve User, and retrieve all of a Users data.